[iOS #61] Main의 TableView DataSource, Delegate 분리
- DataSource, Delegate 분리
