+++ iOS/SideDish/SideDish/View Controller/MainViewDelegate.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+class MainViewDelegate: NSObject, UITableViewDelegate {
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        let view = tableView.dequeueReusableCell(withIdentifier: HeaderCell.reuseIdentifier)
+        return view
+    }
+
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        return 80
+    }
+
+    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 130
+    }
+}
