+++ iOS/SideDish/SideDish/View Model/Main/MainViewDataSource.swift
@@ -1,0 +1,23 @@
+import UIKit
+
+class MainViewDataSource: NSObject, UITableViewDataSource {
+    // MARK: Rows and Cells
+
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 10
+    }
+
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: ItemCell.reuseIdentifier, for: indexPath)
+
+        return cell
+    }
+
+    // MARK: Sections
+
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 3
+    }
+
+}
+
